// src/generated/openapi/model/outputEnumDTO.ts
/**
 * WTP API Starter
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OutputEnumDTO { 
    name: string;
    values: Array<string>;
    isIncluded?: boolean;
}