// src/generated/openapi/model/outputModelDTO.ts
/**
 * WTP API Starter
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OutputFieldDTO } from './outputFieldDTO';


export interface OutputModelDTO { 
    name: string;
    fields: Array<OutputFieldDTO>;
}